feat: Add objects inlining and separating utilities (#14)
Many of our objects support two flavors of dependencies, we can either
inline the dependency object OR we can define it as a separate entity
and reuse it freely, referencing it by name.

While this feature gives freedom to the end users it comes with a toil
for implementations as they need to take care of the object reference
resolution.

The goal of this PR is to introduce utility functions which would allow
both inlining of referenced objects and separating inlined objects
definitions.