feat: Add fine grained control over reference resolution (#17)
When working with AlertPolicy or SLO we might want to only resolve
(either inline or export) specific dependency objects.

For instance, Nobl9 does not have a separate SLI object and requires the
SLO to have it inlined, but at the same time it expects
AlertNotificaitonTarget to be a separate entity (it maps directly to
Nobl9 AlertMethod object).
Due to these requirements some objects need to be inlined why others
exported before a conversion between OpenSLO and Nobl9 can be achieved.

In order to aid in meeting such more complex requirements we're
introducing `ReferenceConfig` struct which gives fine-grained control to
the `openslosdk` user when both inlining and exporting object
dependencies.