fix: Fix validation for v1 and v2alpha SLO objectives (#18)
The OpenSLO spec defines `op` and `value` as `thresholdMetric` specific
fields.
Initially I chose not to implement any complex validation for these two
fields as SLI can be either inlined, in which case we CAN validate them,
or referenced, in which case we CANNOT do that.

This PR addresses this technical debt, since we already have
capabilities to inline and export referenced objects freely, thanks to
https://github.com/OpenSLO/go-sdk/pull/17, we should harden the
validation when an SLI is inlined.